backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
langgraph
langchain-core
langchain-openai
langchain-anthropic
langchain-google-genai
pydantic>=2.0
pydantic-settings
asyncpg
sqlalchemy[asyncio]
aiosqlite
alembic
redis[hiredis]
langgraph-checkpoint-redis
structlog
python-dotenv
httpx
tenacity
langfuse
numpy
python-jose[cryptography]
